write comments on couple of classes

+++ src/main/java/docSharing/repository/UserRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
         public  User findByEmail(String email);
-        int updateUserEnabledById(Long id, Boolean enabled);
+        Long updateUserEnabledById(Long id, Boolean enabled);
 }
